Clean up enums and configuration
Stop using a mix of environment variables and files for configuration
and instead expect configuration via traitlets like the rest of
JupyterHub. We already have to inject mechanisms to set default values
for those traitlets, so this lets us configure the spawner in the same
way as everything else.

Simplify the complex way of creating enums in favor of a
straightforward str enum with defined values for the lab status.

Change the default admin token path to be under /etc/gafaelfawr so
that we can mount the whole secret on that path, which in turn will
hopefully give us dynamic updates of the token.

Delete more now-uneeded code.