Do all spawning work in start
The original idea to have start return immediately once the lab
start has been initiated doesn't work with JupyterHub's assumptions
about spawners. Its timeouts and error handling expect all of the
work to happen in the start method, and progress must not raise
exceptions or JupyterHub reports uncaught exceptions and breaks its
UI and API.

Follow the design of KubeSpawner and have the start method hold a
copy of its task and do all the progress monitoring and event
creation. The progress method then just waits for that task to
complete, reporting all the events it generates as an iterator.

Do a bit of refactoring to move spawn events to a separate class,
and add severity to the start of the message similar to what
KubeSpawner does.