Improve progress messages and debugging
Be kind to our future selves and return more details in unusual
spawn situations via the event stream. Every bit of debugging
information we can see is helpful. Separate the severity from the
message for spawn events for clearer construction.

Flesh out the docstrings and document which exceptions may be
raised.