Do not reuse labs on conflict
When attempting to create a lab for a user who already has a running
lab, the Nublado lab controller returns a 409 conflict error code.
Previously, the spawner was then returning the URL for that running
lab to JupyterHub instead of spawning a new one, since ideally we
would like to preserve the running lab.

Unfortunately, by the time the spawner start method has been called,
JupyterHub has generated new OpenID Connect credentials for the lab
and the existing running lab cannot talk to JupyterHub. Reusing it
therefore doesn't work; authentication to the Hub fails. We therefore
have to recreate it in order to get a working lab.