Convert to Safir's AsyncMultiQueue
Use the AsyncMultiQueue data structure from Safir to manage the
progress events. This is a richer version of the data structures
that were being used internally.

Rely on the spawner calling end() in its finally block and don't
poll the spawner task for completion inside the progress iterator.
The Safir AsyncMultiQueue has a mechanism for explicitly indicating
the end of the data in the queue, which should serve the same
purpose.

Simplify the way that progress events are checked in the test suite
and add an explicit test for reading the events of a spawn after the
spawn has completed.